@@ -16,16 +16,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.5",
+    "@angular/animations": "^14.2.0",
     "@angular/cdk": "13.2.5",
-    "@angular/common": "~13.2.5",
-    "@angular/compiler": "~13.2.5",
-    "@angular/core": "~13.2.5",
-    "@angular/forms": "~13.2.5",
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "@angular/forms": "^14.2.0",
     "@angular/material": "^13.2.5",
-    "@angular/platform-browser": "~13.2.5",
-    "@angular/platform-browser-dynamic": "~13.2.5",
-    "@angular/router": "~13.2.5",
+    "@angular/platform-browser": "^14.2.0",
+    "@angular/platform-browser-dynamic": "^14.2.0",
+    "@angular/router": "^14.2.0",
     "@improbable-eng/grpc-web": "^0.15.0",
     "@ngneat/loadoff": "^2.1.0",
     "@ngneat/transloco": "^3.1.4",
@@ -52,14 +52,14 @@
   },
   "devDependencies": {
     "@angular-builders/jest": "~13.0.3",
-    "@angular-devkit/build-angular": "^13.2.5",
+    "@angular-devkit/build-angular": "^14.2.1",
     "@angular-eslint/builder": "~13.1.0",
     "@angular-eslint/eslint-plugin": "~13.1.0",
     "@angular-eslint/eslint-plugin-template": "~13.1.0",
     "@angular-eslint/schematics": "~13.1.0",
     "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "^13.2.5",
-    "@angular/compiler-cli": "^13.2.5",
+    "@angular/cli": "^14.2.1",
+    "@angular/compiler-cli": "^14.2.0",
     "@ngx-grpc/protoc-gen-ng": "^2.4.1",
     "@types/google-protobuf": "^3.15.5",
     "@types/jasmine": "~3.10.3",
@@ -84,6 +84,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~10.7.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.5.5"
+    "typescript": "~4.8.2"
   }
 }
